+++ src/routes/notebooks/[id]/+page.svelte
@@ -167,6 +167,12 @@
 			allPeople={data.allPeople}
 			categories={data.categories}
 			pickableNotebooks={data.pickableNotebooks}
+			areas={data.areas}
+			workoutMeasures={data.workoutMeasures}
+			slots={data.slots}
+			todos={data.todos}
+			allTodos={data.allTodos}
+			activities={data.activities}
 			bind:newAction
 		/>
 	</section>